feat:Split the schemas structure into request and response
